fix: auth not being accessible by logged in users
